requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==6.0.1
django-allauth==65.14.0
django-widget-tweaks==1.5.1
idna==3.11
jwt==1.4.0
pillow==12.1.0
pycparser==3.0
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
